if.legal/requirements.txt
requests>=2.31.0
httpx>=0.27.0
pandas>=2.2.2
pypdf>=4.2.0
chromadb>=0.5.3
beautifulsoup4>=4.12.3
python-slugify>=8.0.4

